summon adjusting loot and required items (#2801)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Does the following:

- makes all summons consistent requiring 2 of the BASIC materials needed
for summoning.
- increases health and damage of behemoth, dryad, and void dragon
- fixes broken drops on some summons: behemoth, watcher, fiend

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Most of this was oversights being fixed.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/85a79a09-4333-437d-8c8f-1fc4d599cba3


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->