Adds being blind as a starting flaw. (#2827)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds blindness as a selectable flaw, it also makes you spawn with a
wooden staff that will allow you to examine things.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Allows players to play the gimmick of their character being blind
without the need of admin intervention, it also makes selecting a random
flaw more risky.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
**The blind flaw**
![Screenshot 2025-05-08
010759](https://github.com/user-attachments/assets/605cc350-bcdc-4ef1-86d4-9c21dd534e53)

Sends a message when you examine something with the wooden staff.
![Screenshot 2025-05-08
012909](https://github.com/user-attachments/assets/dc963f2c-91e4-4d2b-bbc6-34fcd046e4ca)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->