Fix for bandit liches (#2853)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adds bandit to the list of roles that cannot roll lich.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
As far as I know, bandits aren't supposed to be able to be liches - they
can't be any other type of antag, so the omission of bandits from the
exclusion list seems to just be an accident, probably dating back to
when bandits were more formal antagonists rather than a selectable role.
Additionally, this is an issue that seems to crop up alarmingly
frequently in rounds - I've seen a number of ahelps by users who've
managed to roll a bandit lich who want to be made into just a bandit,
creating round-start work and depriving the round of a proper antagonist
in cases where no alternative can be arranged. This fix would resolve
these issues going forward.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/aab9d8c8-4859-4434-a423-6bfefab754c2)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->