Graggarites can drink now small portions of blood to quench thirst. (#2804)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PRs adds a way to be slightly more heretical than normal.
Graggarites can drink blood with a single check, 10 units of thirst for
10 blood compared to VL that drinks 45, this is pretty balanced, and it
means you can't drink people to death.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Adds more flavour, RP, and cool things!
Now people can be heretical with blood drinking or just roleplay a blood
bat as a wild-kin.
Well. Hopefully it won't be abused.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/82356416-6e02-4ddf-bfc3-dbe3644d0298)
no vomiting.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

---------

Co-authored-by: Sutures <101909198+noelle-lavenza@users.noreply.github.com>