random character updates voice type (#2852)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
When the proc to randomize your character is called it adjusts the voice
type to default
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It wont solve previously existing characters (you will have to set
them!!!) but for new characters this means by default you wont have to
in case you forget.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/265e207d-7825-40bc-87b2-4de3e0fffbf2


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->