Scholar Patron Edit (#2851)
Allows Scholar to be either Noc OR Zizo devoted

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Allows Scholars/Archivists to worship either Noc OR Zizo. Literally a 1
line change. (mind blowing, incredible)

## Why It's Good For The Game

It'd be a pretty good role for zizoids, and it makes sense. The Magos
can be Noc or Zizo and this falls in line with that, a scholar would be
a perfect cultist - after all, they've access to tons of knowledge, and
thus there is plenty of chance for them to become corrupted by it.


## Proof of Testing (Required)


![proofoftesting1](https://github.com/user-attachments/assets/7f9fc86c-c22d-472e-ad39-bc5574b69c32)