Fixed move speed not correctly updating (#2870)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes an issue which causes movement speed to not correctly
update when the speed stat is modified.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
In many cases, speed buffs and debuffs are applied simply through
modifying the STASPD value. However, these changes aren't immediately
reflected in the player's movement speed, meaning that these buffs and
debuffs don't function as intended. This PR allows these modifications
to function correctly without any further work, either from developers
or players.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
A video of me running left and right with various speed values using the
effects of Moondust.

https://www.youtube.com/watch?v=-tyQP4vnBts

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->