Adds additional peddler input validation (#2866)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes a bug with peddler input validation, allowing certain
unintended values to be entered. The fix itself is pretty simple, and
copied from AP's fix for the same issue, so my thanks to their team for
getting on top of it.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This issue is one I've noticed come up a few times in recent days and
which has caused some disruption - following the fix, this disruption
should no longer be possible.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/a7e4d4df-e2d3-466e-bacf-bac88e3bd904)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->