[Experimental] Initial support for non-despawning summonable mobs (#2867)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adds (very basic) support for non-despawning summonable mobs for
the BYOS event, potential future events, and perhaps even dungeons and
other PVE content down the line.

All that has changed is summonable mobs now have a "despawn_on_idle"
variable, defaulting to true, which determines whether the mob should
despawn when its AI enters the idle state. No further changes have been
made.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The change potentially allows for future mapping opportunities - in the
form of being able to place summonable mobs that don't despawn - without
needing to create new versions of them, through setting the
despawn_on_idle variable. It will be of particular utility in the
upcoming BYOS event, which this was implemented for, but could prove
useful later.

That said, it may need some further refinement (eg. the ability to turn
off drops for these mobs). This hasn't been implemented yet as it wasn't
requested for the upcoming event.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Not much to show off here, but the changes build fine.

![image](https://github.com/user-attachments/assets/f80c928d-f254-409b-b3b5-0ee3edadd232)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->