Removes the game forcing your skin tone if you choose Grenzelhoft mercs (#2876)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Removes the Grenzelhoft whitewashing forcefield, aka a few lines of code
that will force any humen to be white, with approval from a lore
maintainer. Gone are the days of our few Ebon players choosing Grenzel
and getting suddenly whitewashed.

## Why It's Good For The Game

No real reason to keep it, all it does is ruin certain players immersion
when they choose Grenzel merc and are suddenly a pasty white man because
they don't know that the Grenzels code forces everyone white


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/4ed5b430-01ff-40f2-830f-59398827499c)

![image](https://github.com/user-attachments/assets/50673b79-6c54-428b-879b-f68c7e685aa3)