berry bush and apple tree: 10 mins to mature, 6 min produce time (#2861)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
see title
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
people have apples everywhere by the end of the round, these are the two
produce crops that do not delete on creation and as thus it may make
more sense if they grow a bit slower to make up for it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
var edits (code builds)
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->