Rewrite roguetownprocs.dm to fix issues and improve code quality. (#2826)
## About The Pull Request
`roguetownprocs.dm` is by far the most infamous file in all Roguetown
forks, given how a lot of the code is poorly-written, redundant, in
violation of basic object-oriented programming principles,
poorly-organised, redundant, buggy, outright broken, and redundant. This
rewrites nearly all of it, fixing a few bugs in the process.

- `checkdefense()` is now split into three procs, `checkdefense()`,
`checkparry()`, and `checkdodge()`. In terms of functionality they
should be the same, the code is just way easier to understand (IMO) now.
- You can no longer dodge into turfs with dense objects.
- Simplemob `defprob` now actually applies; as-written it would never
ever work because lol bad code.

## Why It's Good For The Game
Code quality good, bugs bad.

## Proof of Testing (Required)
I have not tested it yet.