Fix for left arm spawn on reapplied game preferences (#2855)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes a longstanding issue in which reapplying a player's game
preferences would cause a phantom left arm to spawn, due to a flaw in
the logic which replaces their arms to account for the missing arm
character flaws. The fix is simply calling qdel() on the created arm -
this was already being done for the right arm, but seems to have been
omitted for the left due to an oversight.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This fix saves admins time when performing a common operation (often
applied multiple times per round); in particular it will, I think, no
longer be necessary to aghost as part of the process to delete the
phantom limb spawn. It also prevents players from seeing left arms
popping into and out of existence and potentially impacting immersion.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
I won't bore you with a screenshot of me with no arm on the ground, and
then another screenshot of me with no arm on the ground.

![image](https://github.com/user-attachments/assets/48e7dae3-7fa2-42d9-923d-8c355a5e2896)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->