Allows leaving via carriage and more... (#2656)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
What it does:

- Puts a door item on top of the carriage's door on new you can click on
and examine.
- Clicking it gives you a prompt to leave
- Allows you to leave by clicking on the carriage, you enter after 5
seconds.
- Cannot leave if you are considered important jobs `(ruler, prince,
consor, priest) `
- Checks for special items that should never be allowed to leave (master
key,crown of rockhill, priest staff)
- Also checks the inventories to make sure you're not hiding it!
- leaving deletes the character and removes it from all the lists,
leaves you a spirit (This was the easiest way to deal with a cooldown to
keep it consistent)
- opens up the slot you took up and subclass (thanks Azure Peak)
- Logs who left and with what and what job to admins (So if some idiot
abuses this you can call them out on it)

pls don't abuse this gamers :'(
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People want to leave the round in a way that doesn't mean perma sleeping
or killing themselves. This allows them to. I may need to adjust things
going forward depending on what people do with this but I think we can
trust most people to not abuse this. Especially considering how many
people hate the underworld it would at least filter some people from
trying to game this...

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/4928354a-f6eb-43b2-b2cc-fc3fbe8d057b


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->