Skeletons can tell if somebody is a zizoid (#2872)
## About The Pull Request

Skeletons will now be able to tell who is and who is not a zizoid.

## Why It's Good For The Game

Because it leads to nonstop "friendly fire" incidents. Because setting
wise, this seems appropriate. At the very least, if not this, then they
should get an examine text for people in the undead faction such as
apostates.

## Proof of Testing (Required)


![eIFzyQj](https://github.com/user-attachments/assets/ca8eaa17-eb67-43e3-958d-f5ae85859e09)