Allows humans to vomit via emotes (#2873)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
adds vomit emote
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Funny. Also there could be times where someone wants to RP vomitting and
now they could add this into their actions. It still goes through the
vomit emote so you still get all the negative effects that come with
this (It is not an aesthetic action!!) so if you spam this or abuse it
it **will** leave you completely open and defenseless. So don't go
abusing it to puke on someone's floor or they'll get several seconds to
beat your ass for free!

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/398d1924-cbfc-4488-99da-878dc5d8618e


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->