Don't attempt to show hover over text for qdel'd turf/mob/atoms (#2856)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR fixes an issue that mostly affects high ping users when hovering
over items that no longer exist on the server, most commonly collected
items.

There is a rare chance that it'll attempt to render the offset position
for a qdel'd item, which has been moved so far off screen that'll cause
the client viewport to turn black, and require a reconnect to fix.

<img width="1388" alt="hmm"
src="https://github.com/user-attachments/assets/19c5cbaf-9a71-43e3-a9d4-b49f63ebe789"
/>

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

This fixes a techincal bug by adding additional checks to the hover over
logic to avoid rendering items for qdel'd objects.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)

I've recorded a video of this bug in action, and the fix being shown
working by logging in chat.
```dm
/atom/proc/handle_mouseover(location, control, params)
	var/mob/p = usr
	if(QDELETED(src))
		to_chat(p, span_notice("attempted to render deleted object [name]"))
		return FALSE
	...
```


https://github.com/user-attachments/assets/a0fbce00-38ae-48e9-a3ea-c2e6f81e17e5

By using the same checks when collecting items as for hovering, it'll
avoid rendering hover over items that'll be invalid as the object no
longer exists on the server, but still exists on the client due to ping
latency.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->