Jam's Lowtown map changes + Goblin Encampment Revamp
A fairly thorough adjustment of the lowtown area, rearranging buildings, changing the landmass, the shape of the vanguard fort, and dozens of small changes. The intention is to add more points of interest and to position things such to direct player traffic in a way that is more conducive to fun encounters - making the soilery less secluded and including routes around it and points of interest behind it so that there are reasons for players to happen past the soilery without soley having the intention to go there and nowhere else, bringing the faces of the blacksmith and cheesemakers closer  so they face a central town square for people to encounter eachother, including an outdoor training area for the vanguard to spar in that passerbys can observe, etc.