Various fixes
There was a shack missing a roof. I also gave it a basement for people do do evil in. Also fixed a lot of doors and windows that had walls in them that I hadn't noticed. I also realised that Towner Doors only give you keys for a single door, so I replaced the backdoors I added to the cheeser's and smith's with latch doors that are locked by default. A myriad of other fixes around things around doors, windows and stairs.