Fixed player panel not working correctly for ghosts (#2890)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR corrects an issue causing admins to be unable to open the Player
Panel for players who are currently ghosts. The issue is a simple
indentation issue - the logic which actually opens the player panel is
indented one level too much, causing it to be conditional on whether the
targeted mob is living.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes life easier for admins - the player panel is quite important to be
able to access for ghosts; for example, to return them to the lobby, or
to open their logs (for example, to investigate a dead player's claims
to have been attacked in a rule-breaking way).

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Myself as a ghost (leftmost part of the image), with my player panel
open. Player panel was opened after dying and exiting my body.

![image](https://github.com/user-attachments/assets/6d84633e-0e93-4e9b-aa63-1998d4927a9b)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->