Damage Threshold on limbs before Crit (#2729)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR ports part of
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/350 and
creates a damage threshold that has to be met for limbs before a crit
can occur; Unless the target is either buckled, or laying down.
It also increases bleedrate for minor wounds by a factor of 2.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The idea behind this PR/Port is to reduce NPC Mobs critting people on
one hit, as well as reduce 1 hit fight ending moments.
Having an epic fight with an opponent; only for the first hit to
immediately crit an artery? Not exactly fun nor an engaging system.
This also makes CON a more important stat as the threshold for limbs
will change based on max health; meaning more con, more damage needed
per limb to crit.
Minor wounds are also more punishing bloodloss wise.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->