Mayor minor rebalancing (#2842)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
I have been noticing a lot of flaws recently by playing Mayor almost all
the time.
This was discussed by Ezra on discord and most of this changes are part
of the sequences of changes i will do to improve this and other roles.
Initial PR is to be merged somewhat quickly to fix key issues so the
mayor can do their job because holy shit man how i am supposed to be the
one in charge of the watch if i can't even enter the damn watch itself.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Mayor better and good and nice drip.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/2a1f85ef-81fc-40eb-967d-68fedb986eb8)

![image](https://github.com/user-attachments/assets/1dab14eb-a34a-4f02-9093-984f2822de58)

![image](https://github.com/user-attachments/assets/3478d1e2-41c1-4afc-977e-d70d7a6c9275)

![image](https://github.com/user-attachments/assets/4fb5068e-6bf4-4160-9801-6d394b2699d3)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->