Adventurers back as migrant wave (#2822)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

popular content dont touch things that do work perfectly
part of bog ressurection

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

something 

## Proof of Testing (Required)

something 


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->