Fix for sewn armour not graphically updating when repaired & examine text not correctly identifying broken items (#2897)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes two issues:
1) Sewn items, when repaired, did not correctly update their sprites
from the broken state.
2) When examined, broken items' examine text would not necessarily say
that the item was broken (ie. unable to provide protection) if its break
integrity threshold was greater than 1% of the maximum integrity.

This is achieved by replacing direct mutations of the obj_integrity
value with a call to a new mend_damage proc, which handles modifying the
integrity, enforcing the maximum integrity, and updating the object's
graphics in the event that it's mended from a broken state. This proc
takes a second argument which defines whether it will be allowed to mend
a broken item or not.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The "broken" sprites for equipment are - deliberately - unsatisfying and
unappealing to look at, and remaining in that state makes it difficult
to visually identify whether an item has been broken and subsequently
mended, or is still broken. This fix makes it easier to tell at a glance
whether the armour will still provide protection or not.

Additionally, examine text saying that an item is severely damaged when
it's actually broken and can provide no protection is misleading; this
fix makes the text more reflective of the item's actual ability to
protect, reducing possible player confusion.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Myself wearing chain chausses and a padded gambeson, which I'll be
damaging with the Crimson Fang.

![image](https://github.com/user-attachments/assets/64dc7fcf-77f3-46a2-91f0-431da68beb4a)

The padded gambeson and chain chausses on a table, in the broken state
and showing the appropriate graphics.

![image](https://github.com/user-attachments/assets/3212555c-d24d-4e08-ab17-b8909c7611fb)

Metallic items still can't be repaired with blacksmithing < 4.

![image](https://github.com/user-attachments/assets/cc9df15f-fb7c-4a2b-936c-70d03df6adb9)

Having raised by blacksmithing to 4, the chausses have been repaired to
a functioning (but still low-integrity) state, showing blacksmithing
still works as expected.

![image](https://github.com/user-attachments/assets/303e6e89-6495-4e3a-9e3b-00a4f79a27b7)

Giving myself sewing 2, I've now repaired the padded gambeson to a
functioning state. It now has the unbroken sprite again.

![image](https://github.com/user-attachments/assets/303e6e89-6495-4e3a-9e3b-00a4f79a27b7)

And finally, the object will now display as broken when the broken flag
is set, even if its integrity isn't below 1%:

![image](https://github.com/user-attachments/assets/9b6fd143-db3c-47d6-b60c-79b77d1aca77)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->