Hospitaller ritualist trait & Brother/Sister title (#2898)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes a couple of issues with Hospitallers, specifically:
1) Hospitallers missing the Ritualist trait, meaning they can't make use
of the rune of plague despite having feathers to draw them with. This
seems to have been a simple oversight when porting the rituals from AP,
since AP does not - as I understand - have Hospitallers.
2) Hospitallers having no defined title, while almost all other church
roles have "Brother" and "Sister" as titles. Again, this looks likely to
have been a simple oversight - but I can remove this if it was a
deliberate choice. Still, I'd argue they should have these titles, to
make them fit in better among the other church roles.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
These two oddities make Hospitallers stand out somewhat among Church
roles as not entirely fitting in, due to lacking a capability of other
roles and also lacking any sort of title, leaving it unclear, in-game,
whether they should be referred to as "Brother"/"Sister" or not. This PR
amends these two issues, hopefully making Hospitallers feel a little
more like they fit in among the rest of the church.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Screenshot showing a Hospitaller with the ritualist trait, and showing
the name including the "Sister" title.

![image](https://github.com/user-attachments/assets/0fd9131c-b270-4b9f-99f9-879ffc7fbc72)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->