Undead swift intent fix (#2906)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR fixes undead antags being able to use the Swift combat intent.

Swift is a very powerful combat intent, allowing a character to attack
(or grab) much more rapidly, at the expense of consuming considerably
more stamina and thus tiring themselves out much more quickly -
something which is likely to be deadly in combat. Skeletons and
vampyres, however, do not consume stamina. For this reason, the Swift
intent is actually removed from those creatures' combat intent options;
however, due to an oversight, if a character has Swift intent selected
when they're turned, they retain it even after it's no longer a valid
intent for them to select.

This PR addresses this problem by checking, on conversion, whether the
character has the Swift intent selected and, if so, forcing their combat
intent back to Feint. Additionally, Swift is removed from the lich's
available combat intents for the same reason; I suspect that the reason
it wasn't already removed was, again, an oversight.

The PR does not modify deadites; this is because deadites don't actually
appear to have the NOSTAMINA trait, and aren't player-controlled anyway.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Ultimately, creatures with TRAIT_NOSTAMINA make Swift intent overpowered
- the intent is mainly held back by its hefty stamina usage, and when
stamina is no longer a concern, it becomes extremely powerful. Indeed,
this was clearly recognised at some point, hence why the intent was
mostly made unavailable to undead - it's only due to the oversight of
forgetting to modify the active intent on conversion that it's possible
to use as an undead at all. This PR simply corrects that mistake -
undead still have the benefit of not requiring stamina, but no longer
benefit from the combat intent which is balanced only by its heightened
stamina consumption.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Myself, with swift intent selected.

![image](https://github.com/user-attachments/assets/96c83daf-233a-4cb5-bde3-ba15cfc52247)

Myself, having been turned into a vampyre via the Traitor Panel. Note
that my combat intent has been forced to Feint.

![image](https://github.com/user-attachments/assets/c214cd76-0688-4f50-a7e1-450178e38764)

Additionally, the Lich can no longer select Swift intent:

![image](https://github.com/user-attachments/assets/12cb9067-2097-46a6-9196-806edd9e3835)

I've also tested with vampyre spawn and animated skeletons, which also
work correctly, and with swift deselection when becoming alich, which
also works correctly; but I'll spare you all the screenshots.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->